--- dojo/db_migrations/0100_disable_features.py
@@ -1,28 +1,0 @@
-# Generated by Django 3.1.8 on 2021-05-08 10:23
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('dojo', '0099_delete_report'),
-    ]
-
-    operations = [
-        migrations.AddField(
-            model_name='system_settings',
-            name='disable_checklists',
-            field=models.BooleanField(default=False, help_text='With this setting turned on, checklists will be disabled in the user interface.', verbose_name='Disable checklists'),
-        ),
-        migrations.AddField(
-            model_name='system_settings',
-            name='disable_credentials',
-            field=models.BooleanField(default=False, help_text='With this setting turned on, credentials will be disabled in the user interface.', verbose_name='Disable credentials'),
-        ),
-        migrations.AddField(
-            model_name='system_settings',
-            name='disable_questionnaires',
-            field=models.BooleanField(default=False, help_text='With this setting turned on, questionnaires will be disabled in the user interface.', verbose_name='Disable questionnaires'),
-        ),
-    ]
