+++ dojo/db_migrations/0100_enable_features.py
@@ -1,0 +1,28 @@
+# Generated by Django 3.1.8 on 2021-05-09 14:12
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('dojo', '0099_delete_report'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='system_settings',
+            name='enable_checklists',
+            field=models.BooleanField(default=True, help_text='With this setting turned off, checklists will be disabled in the user interface.', verbose_name='Enable checklists'),
+        ),
+        migrations.AddField(
+            model_name='system_settings',
+            name='enable_credentials',
+            field=models.BooleanField(default=True, help_text='With this setting turned off, credentials will be disabled in the user interface.', verbose_name='Enable credentials'),
+        ),
+        migrations.AddField(
+            model_name='system_settings',
+            name='enable_questionnaires',
+            field=models.BooleanField(default=True, help_text='With this setting turned off, questionnaires will be disabled in the user interface.', verbose_name='Enable questionnaires'),
+        ),
+    ]
