+++ dojo/templates/base.html
@@ -155,7 +155,7 @@
                                 <li><a href="https://defectdojo.github.io/django-DefectDojo" target="_blank">
                                     <i class="fa fa-question-circle fa-fw"></i> Documentation</a></li>
                                 {{ block.super }}
-                            {% if user.is_staff and not system_settings.disable_questionnaires %}
+                            {% if user.is_staff and system_settings.enable_questionnaires %}
                                 <li><a href="{% url 'questionnaire' %}"> <i class="fa fa-hospital-o fa-fw"> </i> Questionnaires </a></li>
                             {% endif %}
                         {% endblock %}
@@ -319,7 +319,7 @@
                                     </a>
                                 </li>
                             {% endif %}
-                            {% if request.user.is_staff and not system_settings.disable_questionnaires %}
+                            {% if request.user.is_staff and system_settings.enable_questionnaires %}
                                 <li>
                                     <a href="#"><i class="fa fa-hospital-o fa-fw"> </i> <span>Questionnaires</span><span
                                             class="glyphicon arrow"></span></a>
@@ -338,7 +338,7 @@
                                     </a>
                                     <ul class="nav nav-second-level">
                                         {% if request.user.is_superuser %}
-                                        {% if not system_settings.disable_credentials %}
+                                        {% if system_settings.enable_credentials %}
                                         <li><a href="{% url 'cred' %}">Credential Manager </a></li>
                                         {% endif %}
                                         {% if system_settings.enable_jira %}
@@ -511,7 +511,7 @@
                                 <i class="fa fa-history"></i> View Product Tracking Files
                             </a>
                         </li>
-                        {% if not system_settings.disable_credentials %}
+                        {% if system_settings.enable_credentials %}
                         <li role="separator" class="divider"></li>
                         <li role="presentation">
                             <a href="{% url 'new_cred_product' product_tab.product.id %}">
