+++ dojo/templates/dojo/dashboard.html
@@ -127,7 +127,7 @@
         </div>
         <!-- /.col-lg-6 -->
     </div>
-    {% if not system_settings.disable_questionnaires %}
+    {% if system_settings.enable_questionnaires %}
     <div class="row">
         <div class="col-lg-12">
             <div class="panel panel-default">
