+++ dojo/templates/dojo/view_eng.html
@@ -471,7 +471,7 @@
 
   <div id="add_feat" class="panel-body collapse">
 
-    {% if eng.engagement_type == "Interactive" and not system_settings.disable_checklists %}
+    {% if eng.engagement_type == "Interactive" and system_settings.enable_checklists %}
       <div class="panel panel-default">
         <div class="panel-heading">
           <h4> Checklist<span class="pull-right"><a name="collapsible" data-toggle="collapse" href="#add_feat_check_list">
@@ -549,7 +549,7 @@
       </div>
     {% endif %}
 
-    {% if not system_settings.disable_questionnaires %}
+    {% if system_settings.enable_questionnaires %}
     <div class="panel panel-default">
 
       <div class="panel-heading">
@@ -930,7 +930,7 @@
     </div>
 </div>
 {% endif %}
-{% if not system_settings.disable_credentials %}
+{% if system_settings.enable_credentials %}
   <div>
     <div class="panel panel-default-secondary">
       <div class="panel-heading">
