+++ dojo/templates/dojo/view_finding.html
@@ -896,7 +896,7 @@
         </div>
 
         <!--Cred Begin-->
-        {% if  finding.static_finding != True and cred_finding and not system_settings.disable_credentials %}
+        {% if  finding.static_finding != True and cred_finding and system_settings.enable_credentials %}
             <div class="row">
                 <div class="col-md-12" id="cred">
                     <div class="panel panel-default">
