+++ dojo/templates/dojo/view_test.html
@@ -985,7 +985,7 @@
     </div>
 
     <!--Cred Begin-->
-    {% if not system_settings.disable_credentials %}
+    {% if system_settings.enable_credentials %}
     <div class="row">
     <div class="col-md-12" id="cred">
         <div class="panel panel-default">
